+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-06 09:40:43
-// Commit: feat(core): integrate custom telemetry logger in core
+// Updated at: 2026-08-06 11:18:26
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '3377';
-  final String lastMessage = 'feat(core): integrate custom telemetry logger in core';
-  final DateTime timestamp = DateTime.parse('2026-08-06T09:40:43+05:00');
+  final String commitRef = '3382';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-08-06T11:18:26+05:00');
 }